pyproject.toml
[project]
name = "pylintool"
version = "0.1.0"
description = "ANTLR-based Python linter: whitespace fixer, type-check reporter, and code-quality heuristics"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "antlr4-python3-runtime==4.11.1",
    "antlr4-tools>=0.2.1",
    "click>=8.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
]

[project.scripts]
pylintool = "pylintool.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/pylintool"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
